fix(sprint): address review findings (iter 1)
