+++ src/theforge/sprint/runner.py
@@ -86,6 +86,14 @@ from .audit import (
 )
 from .auth_gate import enforce_sprint_auth_readiness
 from .budget import budget_verification_spend, evaluate_budget
+from .carry import (
+    load_sprint_carry_budget_snapshot,
+    previous_run_marker_present as _previous_run_marker_present,
+)
+from .carry import prior_sprint_cost_incomplete as _query_prior_sprint_cost_incomplete
+from .carry import prior_unmeasured_spend_sources as _query_prior_unmeasured_spend_sources
+from .carry import read_prior_sprint_accounting as _query_read_prior_sprint_accounting
+from .carry import read_prior_sprint_audit_cost as _query_read_prior_sprint_audit_cost
 from .ci_checks import PrCheckState, poll_required_checks, required_pr_check_state
 from .collision import (
     batch_group_id,
@@ -122,13 +130,8 @@ from .manifest import (
 from .prior_landing import landing_settled
 from .query import (
     NormalizedDependencyPlan,
-    load_sprint_carry_budget_snapshot,
     normalize_dependency_plan,
 )
-from .query import prior_sprint_cost_incomplete as _query_prior_sprint_cost_incomplete
-from .query import prior_unmeasured_spend_sources as _query_prior_unmeasured_spend_sources
-from .query import read_prior_sprint_accounting as _query_read_prior_sprint_accounting
-from .query import read_prior_sprint_audit_cost as _query_read_prior_sprint_audit_cost
 from .sources import StorySource
 from .state_writer import (
     SPRINT_PHASE_DONE,
@@ -941,7 +944,7 @@ def _read_prior_sprint_cost(project_root: Path, sprint_id: str | None) -> float:
     measured lower bound rather than dropping the prior spend entirely — the
     cost-unknown signal itself is carried by the per-story entries.
     """
-    if not sprint_id or not os.environ.get("FORGE_PREV_RUN_ID"):
+    if not sprint_id or not _previous_run_marker_present():
         return 0.0
     return _query_read_prior_sprint_audit_cost(project_root, sprint_id)
 
@@ -990,7 +993,7 @@ def _read_prior_sprint_accounting(
     if recovered_entries:
         return recovered_cost, earliest_started_at, recovered_entries
 
-    if os.environ.get("FORGE_PREV_RUN_ID"):
+    if _previous_run_marker_present():
         return _read_prior_sprint_cost(project_root, sprint_id), None, {}
 
     return 0.0, None, {}
